# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation pour l'API de gestion des catégories de votre application.",
        "contact": {
            "email": "votre_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/evenements": {
            "get": {
                "summary": "Récupérer toutes les evenements",
                "operationId": "cf17114513864def8e795303a48f7205",
                "responses": {
                    "200": {
                        "description": "Liste des evenements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Ajouter un evenement et ses types de tickets",
                "operationId": "6cc07eed54d06683e60f49ae4d921be1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "description",
                                    "place",
                                    "date_event",
                                    "heure",
                                    "image",
                                    "user_id",
                                    "categorie_id",
                                    "typetickets"
                                ],
                                "properties": {
                                    "titre": {
                                        "description": "Titre de l'événement",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description de l'événement",
                                        "type": "string"
                                    },
                                    "place": {
                                        "description": "Lieu de l'événement",
                                        "type": "string"
                                    },
                                    "date_event": {
                                        "description": "Date de l'événement (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "heure": {
                                        "description": "Heure de l'événement",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Fichier image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "ID de l'utilisateur",
                                        "type": "integer"
                                    },
                                    "categorie_id": {
                                        "description": "ID de la catégorie",
                                        "type": "integer"
                                    },
                                    "typetickets": {
                                        "description": "JSON stringified array of ticket types.",
                                        "type": "string",
                                        "example": "[{\"nom\":\"VIP\",\"prix\":150.0,\"quantite\":10},{\"nom\":\"Standard\",\"prix\":50.0,\"quantite\":100}]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Événement et tickets créés avec succès."
                    },
                    "422": {
                        "description": "Données de validation manquantes ou invalides."
                    },
                    "500": {
                        "description": "Erreur serveur lors de la création."
                    }
                }
            }
        },
...
        "/api/evenements/{id}/validate": {
            "patch": {
                "summary": "Valider un événement (changer son statut à 1)",
                "operationId": "dc6372a069c87467ceef7964f6a550d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement à valider.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Événement validé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Evenement est validé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Événement non trouvé."
                    },
                    "500": {
                        "description": "Erreur serveur lors de la validation."
                    }
                }
            }
        },
        "/api/evenements/{id}": {
            "get": {
                "summary": "Détailles de l'evenement",
                "operationId": "8a580ac0db622f34e0deb0f5fb79364b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Détailles de l'evenement"
                    }
                }
            },
            "post": {
                "summary": "Mettre à jour un événement existant, son image et ses types de tickets",
                "operationId": "e17386b5fae730780ebcb02fa1055f4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'événement à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Champ requis pour simuler la méthode PUT en 'multipart/form-data'",
                                        "type": "string",
                                        "default": "PUT"
                                    },
                                    "titre": {
                                        "description": "Titre de l'événement (Optionnel)",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description de l'événement (Optionnel)",
                                        "type": "string"
                                    },
                                    "place": {
                                        "description": "Lieu de l'événement (Optionnel)",
                                        "type": "string"
                                    },
                                    "date_event": {
                                        "description": "Date de l'événement (YYYY-MM-DD) (Optionnel)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "heure": {
                                        "description": "Heure de l'événement (Optionnel)",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Nouveau Fichier image (Optionnel, remplace l'ancien)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "ID de l'utilisateur (Optionnel)",
                                        "type": "integer"
                                    },
                                    "categorie_id": {
                                        "description": "ID de la catégorie (Optionnel)",
                                        "type": "integer"
                                    },
                                    "typetickets": {
                                        "description": "JSON stringifié (obligatoire pour la mise à jour des tickets). **Tous les anciens tickets seront supprimés et remplacés par cette liste**.",
                                        "type": "string",
                                        "example": "[{\"nom\":\"VIP M.A.J.\",\"prix\":180.0,\"quantite\":5},{\"nom\":\"Standard M.A.J.\",\"prix\":60.0,\"quantite\":150}]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Événement et tickets mis à jour avec succès."
                    },
                    "404": {
                        "description": "Événement non trouvé."
                    },
                    "422": {
                        "description": "Données de validation manquantes ou invalides."
                    },
                    "500": {
                        "description": "Erreur serveur lors de la mise à jour."
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un evenement",
                "operationId": "86e6f8fa5098872c9168eef893ab3876",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Evenement est supprimé"
                    }
                }
            }
        },
...
    },
...
}